Datasets refactoring
Extracted dataset share service
