+++ backend/dataall/api/Objects/Environment/resolvers.py
@@ -21,6 +21,8 @@ from ....utils.naming_convention import (
     NamingConventionPattern,
 )
 
+from dataall.modules.datasets.services.dataset_share_service import DatasetShareService
+
 log = logging.getLogger()
 
 
@@ -391,7 +393,7 @@ def list_shared_with_environment_data_items(
     if not filter:
         filter = {}
     with context.engine.scoped_session() as session:
-        return db.api.Environment.paginated_shared_with_environment_datasets(
+        return DatasetShareService.paginated_shared_with_environment_datasets(
             session=session,
             username=context.username,
             groups=context.groups,
